fix(support-agent): address CodeRabbit review on PR #344
Real bugs:
- Rename `Drafto/Support/Agent-Replied` → `Drafto/Support/Replied` in the
  prompt, fixture, and tests so the suffix allowlist (`Replied`) matches
  the documented label, AND the name fits Zoho's 25-char displayName cap
  (28 chars before, would have hit ENOLABEL same as Needs-Human did).
- Phase-D action handler now hard-fails when Claude returns
  `auto-replied` under Phase D or `filed-issue` under Phase D/E. Before,
  these were silently treated as "no state change required", so a prompt
  regression that bypassed the phase gate would slip through.
- Phase-D prompt: include low-confidence spam (intent=spam,
  confidence<0.85) in the escalation shortcut. Before, those threads
  matched neither step 5 (confidence too low) nor step 6 (intent
  excluded), and steps 7/8 are gated to E/F — leaving them unlabelled
  and looping every 5 minutes.
- Dry-run no longer tees the bundle (which contains customer email
  body, sender, and headers) into LOG_FILE. Per the deferral note from
  PR #335, this PR is exactly the Phase-D rollout that triggers it. The
  bundle still prints to stdout for the operator running interactively.
- `historyFor` in build-bundle.mjs now filters `autoReplies` by
  THREAD_WINDOW_MS at READ time, so `autoReplyCount24h` is a true 24h
  window instead of "count at last bump" (which could include stale
  entries between writes).
- Tighten the summary-line regex to the documented format
  `^thread=<id> action=<x> issue=<n>$`. The looser `^thread=` would have
  accepted any line in Claude's reasoning that started with `thread=`.

Cleanup:
- Extract `parseFlags` to a shared `scripts/lib/parse-flags.mjs` so
  zoho-cli.mjs and state-cli.mjs use the same parser.
- Extract `isMainModule` (using pathToFileURL + realpathSync) to
  `scripts/lib/is-main.mjs`. The naive `import.meta.url === \`file://\${argv[1]}\``
  form fails on symlinks and on paths with spaces / non-ASCII.
- Add a comment header on state-cli.mjs explaining that callers must
  serialize loadState → mutate → saveState (today guaranteed by
  support-agent.sh's PID-file lock).
- Add a comment on `isTerminalSupportLabel` explaining the deliberate
  asymmetry with `assertSupportNamespace` (read tolerant, write strict).
- Drop a misleading "Sanity:" comment in the suffix-allowlist test that
  claimed positive coverage that the test didn't actually exercise.

Two improvements deferred (replied on PR with rationale): switching
`claude -p` to stdin (current argv-based approach is well within macOS's
ARG_MAX for realistic bundles) and replacing
`--dangerously-skip-permissions` with `--allowedTools` (real security
improvement, but requires careful prompt-side coordination — splitting
into a follow-up PR).

All 75 unit tests pass. `pnpm format:check`, `pnpm lint`, `pnpm typecheck`
clean. Live `--dry-run` against the PDF-export fixture verified — no
PII written to LOG_FILE.
